+++ CHANGELOG.md
@@ -1,9 +1,19 @@
-## 1.0.0
+## 1.0.2
 
-- First version
+- FFloatController adds setState function to support external partial refresh of FFloat content
+
+```dart
+controller.setState(() {
+  // update your float ui
+});
+```
 
 ## 1.0.1
 
 - Supports `alignment` configuration in absolute position mode
 
 - Fix FFloatController status monitoring problem
+
+## 1.0.0
+
+- First version